Dynamic step/resolution when fetching data
Background:
Say you have metrics in 60s resolution. When looking at data
over the last few days, this is all well and reasonable. But when you
want to see data over the last 6 months things get slightly out of hand.
In this scenario we're dealing with 259200 datapoints for each series
we're interested in.
When using Whisper/Carbon we most likely storage-schemes to handle this
(something like 1m:21d,15m:5y), and if not, it's reading from disk which
is much faster and has lower latency then querying a remote service.
Most of the time though we'll newer draw that many points. In my grafana
setup it requests a max of 2550 points from graphite-api.

So lets leverage InfluxDBs ability to GROUP BY time.

This commit introduces a new config option called min_data_points.
Setting it to 0 (default) disables dynamic stepping.
Setting it to a value > 0 enables dynamic calculation of the resolution.
The new resolution will always be a multiple of the origin resolution,
and we'll always have at list min_data_points.
Practically this means we'll always have somewhere between
min_data_points and 2xmin_data_points points.